adk/pyproject.toml
[project]
name = "adk-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk[vertexai]>=0.5.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    "google-genai>=1.15.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "litellm>=1.70.2",
    "python-dotenv>=1.1.0",
    "tavily-python>=0.7.2",
]
